refactor(mentor): components/utils 폴더 제거 후 모든 컴포넌트 동일 선상에 배치
