+++ src/components/Header.js
@@ -1,9 +1,9 @@
 import React, { useContext } from "react";
 import "./Header.css";
-import Button from "./utils/Button";
+import Button from "./Button";
 import logoImg from "../assets/logo.svg";
 import { UserContext } from "../contexts/UserContext";
-import ProfileImage from "./utils/ProfileImage";
+import ProfileImage from "./ProfileImage";
 import { Link } from "react-router-dom";
 
 export default function Header() {
