+++ src/components/ProfileImage.js
@@ -1,6 +1,6 @@
 import React from "react";
 import "./ProfileImage.css";
-import profileInit from "../../assets/profile_init.png";
+import profileInit from "../assets/profile_init.png";
 
 export default function ProfileImage({ src, size }) {
   const sizeStyle = `size-${size}`;
