+++ src/components/Search.css
@@ -6,7 +6,7 @@
   border: none;
   outline: none;
   border-radius: 10px;
-  background: #f5f5f5 url("../../assets/search_icon.svg") no-repeat 15px 18px /
+  background: #f5f5f5 url("../assets/search_icon.svg") no-repeat 15px 18px /
     16px;
   font-size: 16px;
   color: #666;
